`Config`: Token Config restructure
