--- src/main/java/dev/piste/vayna/config/tokens/Api.java
@@ -1,22 +1,0 @@
-package dev.piste.vayna.config.tokens;
-
-// GSON CLASS
-@SuppressWarnings("ALL")
-public class Api {
-
-    private String riot;
-    private String henrik;
-    private String topGg;
-
-    public String getRiot() {
-        return riot;
-    }
-
-    public String getHenrik() {
-        return henrik;
-    }
-
-    public String getTopGg() {
-        return topGg;
-    }
-}
