--- src/main/java/dev/piste/vayna/config/tokens/Bot.java
@@ -1,17 +1,0 @@
-package dev.piste.vayna.config.tokens;
-
-// GSON CLASS
-@SuppressWarnings("ALL")
-public class Bot {
-
-    private String vayna;
-    private String development;
-
-    public String getVayna() {
-        return vayna;
-    }
-
-    public String getDevelopment() {
-        return development;
-    }
-}
